Resolving merged conflict
