+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Business/Data/Repository/Interfaces/TestingAndCalibrationService/IUserRepository.cs
@@ -1,11 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using TestingAndCalibrationLabs.Business.Core.Model;
 
 namespace TestingAndCalibrationLabs.Business.Data.Repository.Interfaces
 {
     public interface IUserRepository
     {
         List<string> Get();
+
+        User Get(int id);
+
+        User Get(string userName);
+
+        int Insert(User user, PasswordLogin passwordLogin);
     }
 }
