+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs/bin/Debug/netcoreapp3.1/appsettings.Development.json
@@ -1,0 +1,9 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  }
+}
