+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs/bin/Debug/netcoreapp3.1/appsettings.json
@@ -7,7 +7,7 @@
     }
   },
   "ConnectionStrings": {
-    "IdentityDBLocal": "Server=localhost;Database=TestingAndCalibrationLabs; User ID=sa;Password=admin;"
+    "IdentityDBLocal": "Server=localhost;Database=TestingAndCalibration; User ID=sa;Password=admin;"
   },
 
   "Smtp": {
@@ -25,7 +25,19 @@
     "Subject": "Survey",
     "LogoImage": "https://drive.google.com/uc?export=view&id=1kezy9XtIkZdY7X1MieQuTft1PZHYtrfP",
     "BodyImage": "https://drive.google.com/uc?export=view&id=1XPUvUQiS9BwvxPbYshUWRYTTZ2_G6v8z"
-  },  
+  },
+  "AllowedHosts": "*"
+}
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "ConnectionStrings": {
+    "IdentityDBLocal": "Server=localhost;Database=TestingAndCalibrationLabs;User ID=sa;Password=admin;"
+  },
   "JWT": {
     "ValidAudience": "https://localhost:44337/",
     "ValidIssuer": "https://localhost:44337/",
